+++ js/util.js
@@ -1,0 +1,23 @@
+const getRandomPositiveInteger = (min, max) => {
+  if (min < 0 || max < 0) {
+    return -1;
+  }
+  if (min > max) {
+    const temp = min;
+    [min, max] = [max, min];
+    max = temp;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+const checkLength = (string, maxLength = 140) => string.length <= maxLength;
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {
+  getRandomPositiveInteger,
+  checkLength,
+  getRandomArrayElement
+};
